+++ learning/katas/go/introduction/hello_beam/hello_beam/task-remote-info.yaml
