+++ learning/katas/go/introduction/hello_beam/hello_beam/task.md
