+++ learning/katas/go/introduction/hello_beam/hello_beam_test/task-info.yaml
@@ -17,5 +17,17 @@
 # under the License.
 #
 
-content:
-- Hello Beam
+type: edu
+custom_name: Hello Beam Test
+files:
+- name: pkg/task/task.go
+  visible: true
+- name: test/task_test.go
+  visible: true
+  placeholders:
+  - offset: 1232
+    length: 36
+    placeholder_text: TODO()
+  - offset: 1278
+    length: 12
+    placeholder_text: TODO()
