+++ learning/katas/go/introduction/hello_beam/hello_beam_test/task-remote-info.yaml
