+++ learning/katas/go/introduction/hello_beam/lesson-info.yaml
@@ -17,19 +17,6 @@
 # under the License.
 #
 
-type: edu
-files:
-- name: go.mod
-  visible: false
-- name: cmd/main.go
-  visible: true
-- name: pkg/task/task.go
-  visible: true
-  placeholders:
-  - offset: 920
-    length: 28
-    placeholder_text: TODO()
-- name: go.sum
-  visible: false
-- name: test/task_test.go
-  visible: false
+content:
+- hello_beam
+- hello_beam_test
