+++ learning/katas/go/introduction/hello_beam/lesson-remote-info.yaml
