+++ learning/katas/go/introduction/section-info.yaml
@@ -17,20 +17,5 @@
 # under the License.
 #
 
-type: edu
-files:
-- name: go.mod
-  visible: false
-- name: pkg/task/task.go
-  visible: true
-- name: go.sum
-  visible: false
-- name: test/task_test.go
-  visible: true
-  placeholders:
-  - offset: 1177
-    length: 36
-    placeholder_text: TODO()
-  - offset: 1223
-    length: 12
-    placeholder_text: TODO()
+content:
+- hello_beam
