+++ learning/katas/go/introduction/section-remote-info.yaml
